improve and refactor cron jobs
improve pto policy
