+++ apps/api/src/app.module.ts
@@ -36,6 +36,9 @@ import { UtilService } from './services/util.service';
       },
       resolvers: { JSON: GraphQLJSON },
       context: ({ req, res }: any) => ({ req, res }),
+      includeStacktraceInErrorResponses: process.env.NODE_ENV === 'development',
+      status400ForVariableCoercionErrors: true,
+      autoTransformHttpErrors: true,
     }),
     AuthModule,
     UserModule,
