+++ apps/api/src/modules/pto-tracking/approval-routing/approval-routing.resolver.ts
@@ -1,23 +1,22 @@
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { PtoTrackingService } from './pto-tracking.service';
-import { JwtGraphqlDecorator } from '../../decorators/user.decorator';
+import { JwtGraphqlDecorator } from '../../../decorators/user.decorator';
 import { UseGuards } from '@nestjs/common';
-import { JwtGraphqlGuard } from '../../guards/jwt-graphql.guard';
-import { CreateApprovalRoutingInput } from './dto/create-approval-routing.input';
+import { JwtGraphqlGuard } from '../../../guards/jwt-graphql.guard';
+import { CreateApprovalRoutingInput } from '../dto/create-approval-routing.input';
 
-import { JwtPayload } from '../../types/jwt-payload';
-import { ApprovalRoutingDto } from './dto/approval-routing.dto';
-import { PendingRequestForApprovalDto } from './dto/pending-request-for-approval.dto';
-import { Roles } from '../auth/roles.decorator';
-import { PermissionRoleEnum } from '../../types/permission-role.enum';
-import { UpdateApprovalRoutingInput } from './dto/update-approval-routing.input';
-import { PtoRequestDto } from './pto-request.dto';
-import { PtoRequestStatus } from './dto/pto-employee-request.dto';
+import { JwtPayload } from '../../../types/jwt-payload';
+import { ApprovalRoutingDto } from '../dto/approval-routing.dto';
+import { PendingRequestForApprovalDto } from '../dto/pending-request-for-approval.dto';
+import { Roles } from '../../auth/roles.decorator';
+import { PermissionRoleEnum } from '../../../types/permission-role.enum';
+import { UpdateApprovalRoutingInput } from '../dto/update-approval-routing.input';
+import { PtoRequestStatus } from '../dto/pto-employee-request.dto';
+import { ApprovalRoutingService } from './approval-routing.service';
 
 @UseGuards(JwtGraphqlGuard)
 @Resolver(() => ApprovalRoutingDto)
 export class ApprovalRoutingResolver {
-  constructor(private ptoTrackingService: PtoTrackingService) {}
+  constructor(private approvalRoutingService: ApprovalRoutingService) {}
 
   @Roles(PermissionRoleEnum.MANAGER)
   @Query(() => [PendingRequestForApprovalDto])
@@ -26,7 +25,7 @@ export class ApprovalRoutingResolver {
     @Args('status', { nullable: true, type: () => PtoRequestStatus })
     status?: PtoRequestStatus | null,
   ): Promise<PendingRequestForApprovalDto[]> {
-    return this.ptoTrackingService.getApprovalRequests(
+    return this.approvalRoutingService.getApprovalRequests(
       token.employeeId,
       status,
     );
@@ -37,7 +36,7 @@ export class ApprovalRoutingResolver {
   async approvalRoutings(
     @JwtGraphqlDecorator() token: JwtPayload,
   ): Promise<ApprovalRoutingDto[]> {
-    const result = await this.ptoTrackingService.getApprovalRoutings({
+    const result = await this.approvalRoutingService.getApprovalRoutings({
       companyId: token.companyId,
     });
 
@@ -51,7 +50,7 @@ export class ApprovalRoutingResolver {
     input: CreateApprovalRoutingInput,
     @JwtGraphqlDecorator() token: JwtPayload,
   ): Promise<boolean> {
-    await this.ptoTrackingService.createApprovalRouting({
+    await this.approvalRoutingService.createApprovalRouting({
       config: input,
       companyId: token.companyId,
     });
@@ -65,7 +64,7 @@ export class ApprovalRoutingResolver {
     input: UpdateApprovalRoutingInput,
     @JwtGraphqlDecorator() token: JwtPayload,
   ): Promise<boolean> {
-    await this.ptoTrackingService.updateApprovalRouting({
+    await this.approvalRoutingService.updateApprovalRouting({
       input,
       companyId: token.companyId,
     });
@@ -80,7 +79,7 @@ export class ApprovalRoutingResolver {
     approvalRoutingId: number,
     @JwtGraphqlDecorator() token: JwtPayload,
   ) {
-    await this.ptoTrackingService.assignApprovalRouting(
+    await this.approvalRoutingService.assignApprovalRouting(
       employeeIds,
       approvalRoutingId,
       token.companyId,
@@ -88,41 +87,16 @@ export class ApprovalRoutingResolver {
     return true;
   }
 
-
   @Roles(PermissionRoleEnum.ADMIN)
   @Mutation(() => Boolean)
   async deleteApprovalRouting(
     @Args('id', { type: () => Number }) id: number,
     @JwtGraphqlDecorator() token: JwtPayload,
   ) {
-    await this.ptoTrackingService.deleteApprovalRouting(id, token.companyId);
-    return true;
-  }
-
-  @Roles(PermissionRoleEnum.EMPLOYEE)
-  @Query(() => PtoRequestDto)
-  async getPtoRequestDetails(
-    @Args('requestId', { type: () => Number }) requestId: number,
-    @JwtGraphqlDecorator() token: JwtPayload,
-  ): Promise<PtoRequestDto> {
-    const result = await this.ptoTrackingService.getRequestDetails(
-      requestId,
+    await this.approvalRoutingService.deleteApprovalRouting(
+      id,
       token.companyId,
     );
-
-    return {
-      approvers: result.approvers.map((a) => ({
-        employee: a,
-        status: a.approverStatus,
-        priority: a.priority,
-      })),
-      id: result.id,
-      startDate: result.startDate,
-      endDate: result.endDate,
-      status: result.status,
-      workingDays: result.workingDays,
-      requstedBy: result.requestedBy,
-      ptoCategoryName: result.leaveCategoryName,
-    };
+    return true;
   }
 }
