+++ apps/api/src/modules/pto-tracking/dto/pto-request-approver.dto.ts
@@ -1,0 +1,14 @@
+import { Field, ObjectType } from '@nestjs/graphql';
+import { EmployeeDto } from '../../company/dto/employee.dto';
+
+@ObjectType()
+export class PtoRequestApproverDto {
+  @Field(() => EmployeeDto)
+  employee: EmployeeDto;
+
+  @Field()
+  status: string;
+
+  @Field()
+  priority: number;
+}
