+++ apps/api/src/modules/pto-tracking/leave-accruals/leave-accruals.resolver.ts
@@ -1,12 +1,12 @@
 import { Query, Resolver } from '@nestjs/graphql';
-import { JwtGraphqlDecorator } from '../../decorators/user.decorator';
-import { JwtPayload } from '../../types/jwt-payload';
+import { JwtGraphqlDecorator } from '../../../decorators/user.decorator';
+import { JwtPayload } from '../../../types/jwt-payload';
 import { LeaveAccrualsService } from './leave-accruals.service';
-import { LeaveAccrualsDto } from './dto/leave-accruals.dto';
+import { LeaveAccrualsDto } from '../dto/leave-accruals.dto';
 import { UseGuards } from '@nestjs/common';
-import { JwtGraphqlGuard } from '../../guards/jwt-graphql.guard';
-import { Roles } from '../auth/roles.decorator';
-import { PermissionRoleEnum } from '../../types/permission-role.enum';
+import { JwtGraphqlGuard } from '../../../guards/jwt-graphql.guard';
+import { Roles } from '../../auth/roles.decorator';
+import { PermissionRoleEnum } from '../../../types/permission-role.enum';
 
 @Resolver(() => LeaveAccrualsDto)
 export class LeaveAccrualsResolver {
