+++ apps/api/src/modules/pto-tracking/pto-tracking.service.ts
@@ -4,30 +4,21 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { Database } from '../../database/database.module';
-import { CreateVacationPolicyInput } from './dto/create-vacation-policy.input';
-import { CreateApprovalRoutingInput } from './dto/create-approval-routing.input';
-import { json } from '../../database/json';
 import { LeaveEntitlementDto } from './dto/leave-entitlement.dto';
 import { CreatePtoRequestInput } from './dto/create-pto-request.input';
 import { Transaction } from 'kysely';
 import { DB } from 'kysely-codegen';
-import * as z from 'zod';
 import { jsonArrayFrom, jsonObjectFrom } from 'kysely/helpers/postgres';
-import { VacationPolicyDto } from './dto/vacation-policy.dto';
-import { ApprovalRoutingDto } from './dto/approval-routing.dto';
-import { EmployeeFromDB, mapEmployee } from '../company/util';
-import { UpdateApprovalRoutingInput } from './dto/update-approval-routing.input';
-import { UpdateVacationPolicyInput } from './dto/update-vacation-policy.input';
 import { EventEmitter2 } from '@nestjs/event-emitter';
-import { PtoRequestStatus } from './dto/pto-employee-request.dto';
 import { PaginatedQueryInput } from '../../shared/paginated-query.input';
 import { GetPtoRequestsFilter } from './dto/get-pto-requests.filter';
 import { DateService } from '../../shared/date.service';
-
-type ApprovalConfigJSON = {
-  routing: { approvers: { approverId: number }[] }[];
-  assignedEmployees: number[];
-};
+import {
+  ApproveRejectPTORequestArgs,
+  CreateLeaveCategoryArgs,
+  APPROVAL_ROUTING_JSON_SCHEMA,
+} from './types';
+import { VacationPolicyService } from './vacation-policy.service';
 
 @Injectable()
 export class PtoTrackingService {
@@ -35,6 +26,7 @@ export class PtoTrackingService {
     private readonly database: Database,
     private eventEmitter: EventEmitter2,
     private dateService: DateService,
+    private vacationPolicyService: VacationPolicyService,
   ) {}
 
   getRequestsForEmployee(
@@ -51,8 +43,8 @@ export class PtoTrackingService {
       )
       .orderBy('createdAt desc')
       .select([
-        'LeaveCategory.name as leaveCategoryName',
         'TimeOffRequest.id as id',
+        'LeaveCategory.name as leaveCategoryName',
         'TimeOffRequest.requestedById as requestedById',
         'TimeOffRequest.status as status',
         'TimeOffRequest.workingDays as workingDays',
@@ -66,206 +58,6 @@ export class PtoTrackingService {
         return eb.offset(skip).limit(take);
       })
       .execute();
-  }
-
-  getVacationPolicyForEmployee(employeeId: number, companyId: number) {
-    return this.database.transaction().execute(async (tx) => {
-      const employee = await tx
-        .selectFrom('Employee')
-        .where('Employee.id', '=', employeeId)
-        .where('Employee.companyId', '=', companyId)
-        .select('assignedVacationPolicyId')
-        .executeTakeFirst();
-      if (employee?.assignedVacationPolicyId) {
-        // this is the vacation policy id
-        return this.getPolicyQuery(tx, companyId)
-          .where('VacationPolicy.id', '=', employee.assignedVacationPolicyId)
-          .limit(1)
-          .executeTakeFirstOrThrow();
-      } else {
-        return this.getPolicyQuery(tx, companyId)
-          .where('VacationPolicy.default', '=', true)
-          .limit(1)
-          .executeTakeFirstOrThrow();
-        // return default policy
-      }
-    });
-  }
-
-  private getPolicyQuery(
-    txOrDatabase: Transaction<DB> | Database,
-    companyId: number,
-  ) {
-    return txOrDatabase
-      .selectFrom('VacationPolicy')
-      .where('VacationPolicy.companyId', '=', companyId)
-      .select((eb) => {
-        return [
-          'VacationPolicy.id as id',
-          'VacationPolicy.name as name',
-          'VacationPolicy.description as description',
-          'VacationPolicy.policyDocument as policyDocument',
-          'VacationPolicy.workingDays as workingDays',
-          'VacationPolicy.default as default',
-          'VacationPolicy.archived as archived',
-          jsonArrayFrom(
-            eb
-              .selectFrom('Employee')
-              .whereRef(
-                'Employee.assignedVacationPolicyId',
-                '=',
-                'VacationPolicy.id',
-              )
-              .innerJoin('User', 'User.id', 'Employee.userId')
-              .select([
-                'Employee.id as employeeId',
-                'Employee.companyId as companyId',
-                'User.profilePhotoUrl as profilePhotoUrl',
-                'User.name as fullName',
-                'User.email as email',
-                'Employee.status as status',
-                'Employee.permissionRole as permissionRole',
-              ]),
-          ).as('employees'),
-        ];
-      });
-  }
-
-  getVacationPolicies(args: { companyId: number }) {
-    return this.getPolicyQuery(this.database, args.companyId)
-      .orderBy('default desc')
-      .orderBy('archived asc')
-      .orderBy('createdAt desc')
-      .execute();
-  }
-
-  getVacationPolicy(args: { companyId: number; id: number }) {
-    return this.database.transaction().execute(async (tx) => {
-      const policy: VacationPolicyDto = await tx
-        .selectFrom('VacationPolicy')
-        .where('VacationPolicy.companyId', '=', args.companyId)
-        .where('VacationPolicy.id', '=', args.id)
-        .select((eb) => [
-          'VacationPolicy.id as id',
-          'VacationPolicy.name as name',
-          'VacationPolicy.description as description',
-          'VacationPolicy.policyDocument as policyDocument',
-          'VacationPolicy.workingDays as workingDays',
-          'VacationPolicy.default as default',
-          'VacationPolicy.archived as archived',
-          jsonArrayFrom(
-            eb
-              .selectFrom('LeaveCategory')
-              .where('LeaveCategory.policyId', '=', args.id)
-              .selectAll(),
-          ).as('leaveCategories'),
-          jsonArrayFrom(
-            eb
-              .selectFrom('NationalHoliday')
-              .where('NationalHoliday.vacationPolicyId', '=', args.id)
-              .selectAll(),
-          ).as('publicHolidays'),
-        ])
-        .orderBy('createdAt desc')
-        .limit(1)
-        .executeTakeFirstOrThrow();
-
-      return policy;
-    });
-  }
-
-  createVacationPolicy(data: CreateVacationPolicyInput, companyId: number) {
-    return this.database.transaction().execute(async (tx) => {
-      const hasDefaultPolicy = await tx
-        .selectFrom('VacationPolicy')
-        .where('companyId', '=', companyId)
-        .where('default', '=', true)
-        .select('id')
-        .limit(1)
-        .executeTakeFirst();
-
-      const policy = await tx
-        .insertInto('VacationPolicy')
-        .values({
-          name: data.name,
-          policyDocument: data.policyDocument,
-          description: data.description,
-          companyId,
-          workingDays: data.workingDays,
-          default: !hasDefaultPolicy,
-        })
-        .returningAll()
-        .executeTakeFirstOrThrow();
-
-      const publicHolidays = await tx
-        .insertInto('NationalHoliday')
-        .values(
-          data.publicHolidays.map((holiday) => ({
-            name: holiday.name,
-            date: holiday.date,
-            vacationPolicyId: policy.id,
-          })),
-        )
-        .returningAll()
-        .execute();
-
-      const leaveCategoriesData = data.leaveCategories.map((category) => ({
-        name: category.name,
-        daysAllowed: category.daysAllowed,
-        policyId: policy.id,
-        companyId,
-      }));
-      let leaveCategories: { id: number; name: string; daysAllowed: number }[] =
-        [];
-      if (leaveCategoriesData.length > 0) {
-        leaveCategories = await tx
-          .insertInto('LeaveCategory')
-          .values(
-            data.leaveCategories.map((category) => ({
-              name: category.name,
-              daysAllowed: category.daysAllowed,
-              policyId: policy.id,
-              companyId,
-              accrualType: category.accrualType,
-            })),
-          )
-          .returning(['id', 'name', 'daysAllowed'])
-          .execute();
-      }
-      let employees: {
-        employeeId: number;
-        companyId: number;
-        profilePhotoFilename?: string | null;
-        fullName: string;
-        email: string;
-        status: any;
-        permissionRole: any;
-      }[] = [];
-      if (data.employees.length > 0) {
-        employees = await tx
-          .updateTable('Employee')
-          .innerJoin('User', 'User.id', 'Employee.userId')
-          .innerJoin('Company', 'Company.id', 'Employee.companyId')
-          .set('assignedVacationPolicyId', policy.id)
-          .where('id', 'in', data.employees)
-          .returning([
-            'Employee.id as employeeId',
-            'Employee.companyId as companyId',
-            'User.profilePhotoUrl as profilePhotoUrl',
-            'User.name as fullName',
-            'User.email as email',
-            'Employee.status as status',
-            'Employee.permissionRole as permissionRole',
-          ])
-          .execute();
-      }
-      return {
-        ...policy,
-        leaveCategories,
-        publicHolidays,
-        employees: employees.map(mapEmployee),
-      };
-    });
   }
 
   getLeaveCategories(args: { policyId?: number; companyId: number }) {
@@ -300,116 +92,6 @@ export class PtoTrackingService {
       .executeTakeFirstOrThrow();
   }
 
-  getApprovalRoutings(args: {
-    companyId: number;
-  }): Promise<ApprovalRoutingDto[]> {
-    return this.database.transaction().execute(async (tx) => {
-      const result = await tx
-        .selectFrom('ApprovalRouting')
-        .where('ApprovalRouting.companyId', '=', args.companyId)
-        .orderBy('createdAt desc')
-        .selectAll()
-        .execute();
-
-      const employeeIds = [
-        ...new Set([
-          ...result
-            .map((r) =>
-              (r.config as ApprovalConfigJSON).routing.map((r) => r.approvers),
-            )
-            .flatMap((c) => c.flatMap((c) => c.flatMap((c) => c.approverId))),
-          ...result
-            .map((r) => (r.config as ApprovalConfigJSON).assignedEmployees)
-            .flat(),
-        ]),
-      ];
-
-      let employees: EmployeeFromDB[] = [];
-      if (employeeIds.length) {
-        employees = await tx
-          .selectFrom('Employee')
-          .innerJoin('User', 'User.id', 'Employee.userId')
-          .where('Employee.id', 'in', employeeIds)
-          .select([
-            'User.email as email',
-            'User.name as fullName',
-            'Employee.id as employeeId',
-            'Employee.status as status',
-            'Employee.permissionRole as permissionRole',
-            'User.profilePhotoUrl as profilePhotoUrl',
-          ])
-          .selectAll()
-          .execute();
-      }
-
-      const routings: ApprovalRoutingDto[] = result.map((r) => {
-        return {
-          id: r.id,
-          assignedEmployees: (
-            r.config as ApprovalConfigJSON
-          ).assignedEmployees.map((id) =>
-            mapEmployee(
-              employees.find((e) => e.employeeId === id) as EmployeeFromDB,
-            ),
-          ),
-          approvingLevels: (r.config as ApprovalConfigJSON).routing.map(
-            (routing) => ({
-              approvers: routing.approvers.map((approve) =>
-                mapEmployee(
-                  employees.find(
-                    (e) => e.employeeId === approve.approverId,
-                  ) as EmployeeFromDB,
-                ),
-              ),
-            }),
-          ),
-          config: JSON.stringify(r.config ?? '[]'),
-          name: r.name,
-        };
-      });
-      return routings;
-    });
-  }
-
-  getAssignedApprovalRouting(employeeId: number, companyId: number) {
-    return this.database
-      .selectFrom('Employee')
-      .where('Employee.id', '=', employeeId)
-      .where('Employee.companyId', '=', companyId)
-      .innerJoin(
-        'ApprovalRouting',
-        'Employee.assignedApprovalRoutingId',
-        'ApprovalRouting.id',
-      )
-      .selectAll()
-      .executeTakeFirst();
-  }
-
-  async createApprovalRouting(data: {
-    config: CreateApprovalRoutingInput;
-    companyId: number;
-  }) {
-    await this.database.transaction().execute(async (tx) => {
-      const routing = await tx
-        .insertInto('ApprovalRouting')
-        .values({
-          config: json({
-            routing: data.config.routing,
-            assignedEmployees: data.config.assignedEmployees,
-          }),
-          companyId: data.companyId,
-          name: data.config.name,
-        })
-        .returningAll()
-        .executeTakeFirstOrThrow();
-      await tx
-        .updateTable('Employee')
-        .set('assignedApprovalRoutingId', routing.id)
-        .where('id', 'in', data.config.assignedEmployees)
-        .execute();
-    });
-  }
-
   async getLeaveEntitlementsForEmployee(param: {
     companyId: number;
     employeeId: number;
@@ -441,21 +123,19 @@ export class PtoTrackingService {
     companyId: number,
     employeeId: number,
   ) {
-    const policy = await tx
-      .selectFrom('VacationPolicy')
-      .innerJoin('LeaveCategory', 'LeaveCategory.policyId', 'VacationPolicy.id')
-      .where('LeaveCategory.id', '=', input.categoryId)
-      .where('VacationPolicy.companyId', '=', companyId)
-      .limit(1)
-      .select([
-        'VacationPolicy.archived as archived',
-        'VacationPolicy.id as id',
-        'VacationPolicy.workingDays as workingDays',
-      ])
-      .executeTakeFirstOrThrow();
+    const policy =
+      await this.vacationPolicyService.getVacationPolicyForEmployee(
+        tx,
+        employeeId,
+        companyId,
+      );
+
+    if (!policy) {
+      throw new BadRequestException('No policy found for this employee');
+    }
 
     if (policy.archived) {
-      throw new Error('Policy is archived');
+      throw new BadRequestException('Policy for this employee is archived');
     }
 
     const publicHolidays = await tx
@@ -475,7 +155,7 @@ export class PtoTrackingService {
           input.startDate,
           input.endDate,
           policy.workingDays ?? [],
-          publicHolidays.map((h) => h.date),
+          publicHolidays,
         ),
         status: 'PENDING',
       })
@@ -519,7 +199,8 @@ export class PtoTrackingService {
       .selectFrom('TimeOffRequest')
       .where('TimeOffRequest.id', '=', requestId)
       .where('TimeOffRequest.requestedById', '=', employeeId)
-      .selectAll()
+      .limit(1)
+      .select(['status'])
       .execute();
 
     if (!request) {
@@ -555,7 +236,7 @@ export class PtoTrackingService {
 
     if (approvalRouting) {
       const { config } = approvalRouting;
-      const { routing } = schema.parse(config);
+      const { routing } = APPROVAL_ROUTING_JSON_SCHEMA.parse(config);
 
       // create this payload and create in a single call
       const approvers = routing
@@ -585,46 +266,6 @@ export class PtoTrackingService {
       // todo handle this case
       throw new BadRequestException('No approval routing found');
     }
-  }
-
-  getApprovalRequestsQuery() {
-    return this.database
-      .selectFrom('TimeOffRequestApprover')
-      .innerJoin(
-        'TimeOffRequest',
-        'TimeOffRequest.id',
-        'TimeOffRequestApprover.requestId',
-      )
-      .innerJoin('Employee', 'TimeOffRequest.requestedById', 'Employee.id')
-      .innerJoin('User', 'Employee.userId', 'User.id')
-      .innerJoin(
-        'LeaveCategory',
-        'TimeOffRequest.categoryId',
-        'LeaveCategory.id',
-      )
-      .select([
-        'TimeOffRequest.id as id',
-        'TimeOffRequest.status as status',
-        'TimeOffRequest.startDate as startDate',
-        'TimeOffRequest.endDate as endDate',
-        'TimeOffRequest.workingDays as workingDays',
-        'TimeOffRequest.createdAt as createdAt',
-        'User.name as requestedByName',
-        'LeaveCategory.name as leaveCategoryName',
-      ]);
-  }
-  getApprovalRequests(approverId: number, status?: PtoRequestStatus | null) {
-    return this.getApprovalRequestsQuery()
-      .where('TimeOffRequestApprover.approverId', '=', approverId)
-      .$if(status === PtoRequestStatus.PENDING, (eb) =>
-        eb.where(
-          'TimeOffRequestApprover.status',
-          '=',
-          PtoRequestStatus.PENDING,
-        ),
-      )
-      .orderBy('TimeOffRequest.createdAt desc')
-      .execute();
   }
 
   acceptRejectPTORequest(args: ApproveRejectPTORequestArgs) {
@@ -688,6 +329,7 @@ export class PtoTrackingService {
               timeOfRequestId: id,
             })
             .execute();
+          // todo move to enum this event name
           this.eventEmitter.emit('PTO_REQUEST_FULLY_APPROVED', {
             requestId,
           });
@@ -699,6 +341,7 @@ export class PtoTrackingService {
               a.priority === priority ? a.status === 'APPROVED' : true,
             )
           ) {
+            // todo move to enum this event name
             this.eventEmitter.emit('PTO_REQUEST_LEVEL_APPROVED', {
               requestId,
             });
@@ -707,160 +350,6 @@ export class PtoTrackingService {
       }
 
       return false;
-    });
-  }
-
-  assignApprovalRouting(
-    employeeIds: number[],
-    approvalRoutingId: number,
-    companyId: number,
-  ) {
-    return this.database
-      .updateTable('Employee')
-      .where('id', 'in', employeeIds)
-      .where('companyId', '=', companyId)
-      .set('assignedApprovalRoutingId', approvalRoutingId)
-      .execute();
-  }
-
-  async deleteApprovalRouting(id: number, companyId: number) {
-    return this.database
-      .deleteFrom('ApprovalRouting')
-      .where('id', '=', id)
-      .where('companyId', '=', companyId)
-      .execute();
-  }
-
-  archiveVacationPolicy(param: {
-    companyId: number;
-    id: number;
-    value: boolean;
-  }) {
-    return this.database
-      .updateTable('VacationPolicy')
-      .where('id', '=', param.id)
-      .where('companyId', '=', param.companyId)
-      .where('default', '=', false)
-      .set('archived', param.value)
-      .returning(['id'])
-      .execute();
-  }
-
-  updateApprovalRouting(param: {
-    input: UpdateApprovalRoutingInput;
-    companyId: number;
-  }) {
-    return this.database.transaction().execute(async (tx) => {
-      const currentRouting = await tx
-        .selectFrom('ApprovalRouting')
-        .where('id', '=', param.input.id)
-        .where('companyId', '=', param.companyId)
-        .select('config')
-        .executeTakeFirstOrThrow();
-
-      const config = currentRouting.config as ApprovalConfigJSON;
-      const removedEmployees = config.assignedEmployees.filter(
-        (employeeId) => !param.input.assignedEmployees.includes(employeeId),
-      );
-      const addedEmployees = param.input.assignedEmployees.filter(
-        (employeeId) => !config.assignedEmployees.includes(employeeId),
-      );
-
-      if (removedEmployees.length > 0) {
-        await tx
-          .updateTable('Employee')
-          .set('assignedApprovalRoutingId', null)
-          .where('id', 'in', removedEmployees)
-          .execute();
-      }
-
-      if (addedEmployees.length > 0) {
-        await tx
-          .updateTable('Employee')
-          .set('assignedApprovalRoutingId', param.input.id)
-          .where('id', 'in', addedEmployees)
-          .execute();
-      }
-      await tx
-        .updateTable('ApprovalRouting')
-        .where('id', '=', param.input.id)
-        .where('companyId', '=', param.companyId)
-        .set({
-          config: json({
-            routing: param.input.routing,
-            assignedEmployees: param.input.assignedEmployees,
-          }),
-          name: param.input.name,
-          updatedAt: new Date().toISOString(),
-        })
-        .execute();
-    });
-  }
-
-  async updateVacationPolicy(
-    input: UpdateVacationPolicyInput,
-    companyId: number,
-  ) {
-    await this.database.transaction().execute(async (tx) => {
-      await tx
-        .updateTable('VacationPolicy')
-        .where('id', '=', input.id)
-        .where('companyId', '=', companyId)
-        .set({
-          name: input.name,
-          description: input.description,
-          policyDocument: input.policyDocument,
-        })
-        .execute();
-    });
-  }
-
-  async setAsDefault(id: number, companyId: number) {
-    return this.database.transaction().execute(async (tx) => {
-      const updates = await tx
-        .updateTable('VacationPolicy')
-        .where('companyId', '=', companyId)
-        .where('archived', '=', false)
-        .set('default', false)
-        .returning(['id'])
-        .execute();
-      if (updates.length === 0) {
-        throw new Error('No active policies to set as default');
-      }
-      await tx
-        .updateTable('VacationPolicy')
-        .where('id', '=', id)
-        .where('companyId', '=', companyId)
-        .set('default', true)
-        .execute();
-    });
-  }
-
-  async assignEmployeesToVacationPolicy(
-    param: {
-      employeeIds: number[];
-      removedIds: number[];
-    },
-    policyId: number,
-    companyId: number,
-  ) {
-    await this.database.transaction().execute(async (tx) => {
-      if (param.employeeIds.length > 0) {
-        await tx
-          .updateTable('Employee')
-          .set('assignedVacationPolicyId', policyId)
-          .where('id', 'in', param.employeeIds)
-          .where('companyId', '=', companyId)
-          .execute();
-      }
-      if (param.removedIds.length > 0) {
-        await tx
-          .updateTable('Employee')
-          .set('assignedVacationPolicyId', null)
-          .where('id', 'in', param.removedIds)
-          .where('companyId', '=', companyId)
-          .execute();
-      }
     });
   }
 
@@ -1024,7 +513,7 @@ export class PtoTrackingService {
     startDate: Date,
     endDate: Date,
     workingDays: string[],
-    holidays: Date[],
+    holidays: { date: Date }[],
   ) {
     let workingDaysCount = 0;
     let currentDate = startDate;
@@ -1057,7 +546,7 @@ export class PtoTrackingService {
       if (
         workingDaysInts.includes(currentDate.getDay()) &&
         !holidays.some((holiday) =>
-          this.dateService.createFrom(holiday).isSame(currentDate, 'day'),
+          this.dateService.createFrom(holiday.date).isSame(currentDate, 'day'),
         )
       ) {
         workingDaysCount++;
@@ -1070,24 +559,3 @@ export class PtoTrackingService {
     return workingDaysCount;
   }
 }
-const schema = z.object({
-  routing: z.array(
-    z.object({
-      approvers: z.array(z.object({ approverId: z.number() })),
-    }),
-  ),
-  assignedEmployees: z.array(z.number()),
-});
-
-type CreateLeaveCategoryArgs = {
-  name: string;
-  daysAllowed: number;
-  companyId: number;
-  policyId: number;
-};
-
-type ApproveRejectPTORequestArgs = {
-  requestId: number;
-  accepted: boolean;
-  approverId: number;
-};
