--- apps/api/src/queues/cron-jobs.ts
@@ -1,18 +1,0 @@
-export const CHECK_INS_QUEUE = 'CHECK_INS_QUEUE';
-export const CRON_JOB_FOR_CHECK_INS_UNIQUE_ID =
-  'CRON_JOB_FOR_CHECK_INS_ID_UNIQUE_ID';
-export const CRON_JOB_FOR_CHECK_INS = 'cron-job-for-check-ins';
-
-export const PTO_QUEUE = 'PTO_QUEUE';
-export const PTO_ACCRUALS_CRON_JOB_ID = 'PTO_ACCRUALS_CRON_JOB_ID';
-export const PTO_ACCRUALS_CRON_JOB = 'PTO_ACCRUALS_CRON_JOB';
-
-export const EMPLOYEES_QUEUE = 'EMPLOYEES_QUEUE';
-export const EMPLOYEE_EXPIRED_INVITES_JOB_ID =
-  'EMPLOYEE_EXPIRED_INVITES_JOB_ID';
-export const EMPLOYEE_EXPIRED_INVITES_JOB = 'EMPLOYEE_EXPIRED_INVITES_JOB';
-
-export const EMPLOYEE_PROFILE_IMAGE_GENERATOR_JOB_ID =
-  'EMPLOYEE_PROFILE_IMAGE_GENERATOR_JOB_ID';
-export const EMPLOYEE_PROFILE_IMAGE_GENERATOR_JOB =
-  'EMPLOYEE_PROFILE_IMAGE_GENERATOR_JOB';
