+++ apps/web/src/app/app/dashboard/pto/absence-calendar/components.tsx
@@ -229,7 +229,6 @@ export const CalendarTw = () => {
   const [requestDetailsId, setRequestDetailsId] = React.useState<number | null>(
     null,
   );
-  const [open, setOpen] = React.useState(false);
   return (
     <>
       <div className="flex h-full flex-col">
@@ -281,7 +280,6 @@ export const CalendarTw = () => {
                   }
                   onView={(id) => {
                     setRequestDetailsId(id);
-                    setOpen(true);
                   }}
                 />
               ))}
