+++ apps/web/src/domain/check-in/CurrenctCheckInView.tsx
@@ -66,7 +66,6 @@ export const CurrentCheckInView = () => {
   const hasPendingCheckIns =
     !loading && (data?.getPendingCheckins.length ?? 0) > 0;
 
-
   return (
     <div className="flex flex-col">
       {hasPendingCheckIns ? (
